Move interaction kernel to interaction dir
